# libs/api/domain/auth/src/lib/auth.controller.ts
import {
  Controller,
  Post,
  Body,
  Get,
  UseGuards,
  Request,
} from '@nestjs/common';
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBody,
  // ApiBearerAuth,
} from '@nestjs/swagger';
import { AuthService } from './auth.service';
import { RegisterDto } from './dto/register.dto';
// import { LoginDto } from './dto/login.dto';
import { AuthResponseDto } from './dto/auth-response.dto';
// import { UserResponseDto } from '../users/dto/user-response.dto';
// import { JwtAuthGuard } from './guards/jwt-auth.guard';

@ApiTags('auth')
@Controller('auth')
export class AuthController {
  constructor(private readonly authService: AuthService) {}

  @Post('register')
  @ApiOperation({ summary: 'Register a new user' })
  @ApiBody({ type: RegisterDto, description: 'User registration details' })
  @ApiResponse({
    status: 201,
    description: 'User registered successfully',
    type: AuthResponseDto,
  })
  @ApiResponse({ status: 400, description: 'Bad request' })
  async register(@Body() registerDto: RegisterDto): Promise<AuthResponseDto> {
    return this.authService.register(registerDto);
  }

  // @Post('login')
  // @ApiOperation({ summary: 'Login with email and password' })
  // @ApiResponse({
  //   status: 200,
  //   description: 'Login successful',
  //   type: AuthResponseDto,
  // })
  // @ApiResponse({ status: 401, description: 'Unauthorized' })
  // async login(@Body() loginDto: LoginDto): Promise<AuthResponseDto> {
  //   return this.authService.login(loginDto);
  // }

  // @Get('me')
  // @UseGuards(JwtAuthGuard)
  // @ApiBearerAuth()
  // @ApiOperation({ summary: 'Get current user profile' })
  // @ApiResponse({
  //   status: 200,
  //   description: 'Returns the current user profile',
  //   type: UserResponseDto,
  // })
  // async getProfile(@Request() req): Promise<UserResponseDto> {
  //   return this.authService.getProfile(req.user.id);
  // }
}